requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
redis==4.6.0  # для кэширования
celery==5.3.4  # для асинхронных задач
Flask-Caching==2.0.2